Create tests for Distance Between Stops
